feat: finish contribution module
