+++ TurboBudget/Core/Navigation/AppDestination.swift
@@ -12,6 +12,7 @@ import PaywallModule
 import TransactionModule
 import SubscriptionModule
 import SavingsPlanModule
+import ContributionModule
 
 extension AppDestination {
